doc-level-test-set/requirements.txt
blis==0.7.5
boto3==1.20.24
botocore==1.23.24
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cymem==2.0.6
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
idna==3.3
importlib-metadata==4.8.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
neuralcoref==4.0
nltk==3.2.5
numpy==1.21.5
packaging==21.3
pathy==0.6.1
preshed==3.0.6
psutil==5.9.0
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
regex==2021.11.10
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stanfordcorenlp==3.9.1.1
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
ufal.nametag==1.1.2.1
ufal.udpipe==1.2.0.3
urllib3==1.26.7
wasabi==0.9.0
zipp==3.6.0
